Relocate decorators into distinct namespace, update signatures
